# Memoize `verify_proof` host function