--- a/rusk-abi/Cargo.toml
+++ b/rusk-abi/Cargo.toml
@@ -28,6 +28,8 @@ piecrust = { version = "0.17.0", optional = true }
 # These are patches since these crates don't seem to like semver.
 rkyv = { version = "=0.7.39", default-features = false, features = ["size_32"] }
 
+lru = "0.12"
+
 [dev-dependencies]
 rand_core = { version = "0.6", default-features = false, features = ["getrandom"] }
 once_cell = "1.15"