We memoize the `verify_proof` host function, allowing proof verification to happen only once per equivalent proof being verified in the context of transaction execution.

This is achieved by using a static LruCache, with a configurable maximum limit of entries. These entries are key-value pairs, from the hash of the parameters used to call the proof verification to the outcome of said verification, which allows us to skip verification if we've already performed it once.

Resolves #1228