@ureeves 2024-03-21
> LVGTM
> 
> However Consensus will not benefit at all from this memoization, unless the rusk::preverify use this mechanism too.

Just made some changes that make this happen. Its a little slower because of `JubJubScalar` serialization, but should be perfectly acceptable.