@herr-seppia 2024-03-21
@ureeves Wait.... does this means 250TPS???