@HDauven 2024-03-22
On the Pi 5 I'm seeing the following results:

| TXs in Block | Without Cache (ms) | With Cache (ms) | Improvement |
|------------|-----------------|--------------|-------------|
| 1            | 64                  | 30              | 52.9%    |
| 2            | 104                | 37              | 64.3%    |
| 3            | 147                | 46              | 68.6%    |
| 4            | 192                | 55              | 71.3%    |
| 5            | 234                | 64              | 72.6%    |
| 10          | 444                | 107            | 75.9%    |
| 20          | 865                | 193            | 77.7%    |
| 30          | 1295              | 281            | 78.3%    |
| 40          | 2111              | 456            | 78.4%    |
| 50          | 2563              | 546            | 78.7%    |
|100         | 4279              | 890            | 79.2%    |