@HDauven 2024-03-22
DO CPU-optimized 4 CPU 8 GB:

| TXs in Block | Without Cache (ms) | With Cache (ms) | Improvement |
|------------|-----------------|--------------|-------------|
| 1            | 37                  | 21              | 44.9%    |
| 2            | 60                | 26              | 56.0%    |
| 3            | 81                | 32              | 60.6%    |
| 4            | 102                | 37              | 63.5%    |
| 5            | 125                | 44              | 64.8%    |
| 10          | 234                | 73            | 68.8%    |
| 20          | 453                | 130            | 71.3%    |
| 30          | 672              | 187            | 72.2%    |
| 40          | 891              | 245            | 72.5%    |
| 50          | 1111              | 303            | 72.7%    |
|100         | 2207              | 592            | 73.2%    |